requirements.txt
appdirs==1.4.4
biopython==1.79
black==21.7b0
click==8.0.1
flake8==3.9.2
# Editable Git install with no remote (futils==0.0.1)
-e /data/Imperial/src/futils
isort==5.9.2
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.1
pathspec==0.9.0
psutil==5.8.0
pycodestyle==2.7.0
pyflakes==2.3.1
regex==2021.7.6
tomli==1.0.4
tqdm==4.61.2
